Add files via upload
This was an attempt at a random forest regression/classification model, but unfortunately, it turned out that the data was not viable for models of these types or really any model at all. I found that this was a project that I learned more on than anything else I have ever done and it was an incredibly valuable experience.